feat: refactor FII routes and add the Notion
