--- src/routers/index.ts
@@ -1,3 +1,0 @@
-import FiisRoutes from "./fiis";
-
-export { FiisRoutes };
