+++ src/routes/index.ts
@@ -1,0 +1,4 @@
+import FiisRoutes from "./fiis";
+import NotionRoutes from "./notion";
+
+export { FiisRoutes, NotionRoutes };
